Quiet compiler warning in dbc_t::spec_by_idx().
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@14796 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/dbc/sc_const_data.cpp
@@ -2192,7 +2192,7 @@ specialization_e dbc_t::spec_by_idx( const player_e c, unsigned idx ) const
 {
   int cid = util::class_id( c );
 
-  if ( ( cid <= 0 ) || ( cid >= specialization_max_class() ) || ( idx >= MAX_SPECS_PER_CLASS ) )
+  if ( ( cid <= 0 ) || ( cid >= static_cast<int>( specialization_max_class() ) ) || ( idx >= MAX_SPECS_PER_CLASS ) )
   {
     return SPEC_NONE;
   }
